fixed problem with get() override. On char quit, it would initiate item extraction from the character object. each of these is finalized with an environment.get() which would subsequently save the char, which.. would end up saving you with 0 items.. oops.
Changed save file format from json to pickle. The saved data is the json string, which is pickled.

On my test setup this has improved whole game state save time by 3s or more. Pickling the json doesnt care or worry about the classes that json has serialized, and affords a very small amount of security with the files themselves. Bonus, it writes to file faster.

To switch between the two flip the save format option in settings.py change the Pickle def in SAVE_FORMAT from True to False, and JSON from False to True